Move JobsBuilder under ConsumerGroups namespace
Move `Karafka::Processing::JobsBuilder` to
`Karafka::Processing::ConsumerGroups::JobsBuilder` and
`Karafka::Pro::Processing::JobsBuilder` to
`Karafka::Pro::Processing::ConsumerGroups::JobsBuilder`.

The jobs builder only creates consumer-group-specific jobs (Consume,
Eofed, Revoked, Shutdown, Idle, Periodic). Moving it under the
ConsumerGroups namespace prepares the processing layer for a parallel
ShareGroups::JobsBuilder once KIP-932 lands.

Internal-only move, no backward-compat alias needed.
