+++ spec/lib/karafka/pro/processing/consumer_groups/jobs_builder_spec.rb
