+++ spec/lib/karafka/processing/consumer_groups/jobs_builder_spec.rb
